+++ CorruptionTooltips.lua
@@ -1,1 +1,2 @@
 CorruptionTooltips = LibStub("AceAddon-3.0"):NewAddon("CorruptionTooltips")
+CorruptionTooltips.frame = CreateFrame("Frame", "CorruptionTooltipsFrame", UIParent)