+++ Modules/Icons.lua
@@ -1,7 +1,7 @@
 local Addon = LibStub("AceAddon-3.0"):GetAddon("CorruptionTooltips")
 local Module = Addon:NewModule("Icons", "AceEvent-3.0", "AceHook-3.0")
 
-local Config, Scanner, debounce
+local Config, Scanner
 
 function Module:OnInitialize()
     Config = Addon:GetModule("Config")
@@ -59,9 +59,7 @@ local function SetPaperDollCorruption(button, unit)
         local itemLink = GetInventoryItemLink(unit, slotId)
         if (itemLink) then
             if IsCorruptedItem(itemLink) then
-                local _, icon = Scanner:GetCorruptionByItemLink(itemLink)
-                CreateCorruptionIcon(button, icon)
-                AddNzothIconOverlay(button)
+                Module:ApplyIcon(button, itemLink)
             else
                 if button.corruption then
                     button.corruption:Hide()
@@ -98,11 +96,7 @@ local function SetEquipmentFlyoutCorruption(button)
     end;
 
     local itemLink = C_Item.GetItemLink(itemLoc)
-    if IsCorruptedItem(itemLink) then
-        local _, icon = Scanner:GetCorruptionByItemLink(itemLink)
-        CreateCorruptionIcon(button, icon)
-        AddNzothIconOverlay(button)
-    end
+    Module:ApplyIcon(button, itemLink)
 end
 
 local function SetContainerButtonCorruption(button, bag)
@@ -116,14 +110,8 @@ local function SetContainerButtonCorruption(button, bag)
         return
     end
 
-    -- item:ContinueOnItemLoad(function()
-        local itemLink = GetContainerItemLink(bag, slot)
-        if IsCorruptedItem(itemLink) then
-            local _, icon = Scanner:GetCorruptionByItemLink(itemLink)
-            CreateCorruptionIcon(button, icon)
-            AddNzothIconOverlay(button)
-        end
-    -- end)
+    local itemLink = GetContainerItemLink(bag, slot)
+    Module:ApplyIcon(button, itemLink)
 end
 
 local function SetLootCorruption(index)
@@ -136,10 +124,7 @@ local function SetLootCorruption(index)
     local slot = (numLootToShow * (LootFrame.page - 1)) + index
     if (LootSlotHasItem(slot)) then
         local itemLink = GetLootSlotLink(slot)
-        if itemLink and IsCorruptedItem(itemLink) then
-            local _, icon = Scanner:GetCorruptionByItemLink(itemLink)
-            CreateCorruptionIcon(button, icon)
-        end
+        Module:ApplyIcon(button, itemLink)
     end
 end
 
@@ -181,79 +166,13 @@ function Module:OnEnable()
         self:SecureHook("LootFrame_UpdateButton", function(index)
             SetLootCorruption(index)
         end)
-        if IsAddOnLoaded("AdiBags") then
-            LibStub('ABEvent-1.0').RegisterMessage("CorruptionTooltips", "AdiBags_BagOpened", CT_AdiBags_UpdateAfter)
-            LibStub('ABEvent-1.0').RegisterMessage("CorruptionTooltips", "AdiBags_ForceFullLayout", CT_AdiBags_UpdateAfter)
-        end
-        if IsAddOnLoaded("Bagnon") then
-            self:SecureHook(Bagnon.Item, "Update", CT_Bagnon_Update)
-            self:SecureHook(Bagnon.Item, "UpdateBorder", CT_Bagnon_UpdateBorder)
-        end
     end
 end
 
-function CT_AdiBagsItemButton_CIMIUpdateIcon(button)
-    if not button then return end
-    local bag, slot = button.bag, button.slot
-    -- need to catch 0, 0 and 100, 0 here because the bank frame doesn't
-    -- load everything immediately, so the OnUpdate needs to run until those frames are opened.
-    if (bag == nil) or (slot == nil) or (bag == 0 and slot == 0) or (bag == 100 and slot == 0) then return end
-    local itemLoc = ItemLocation:CreateFromBagAndSlot(bag, slot)
-
-    if (not itemLoc:IsValid()) then return end
-
-    local itemLink = C_Item.GetItemLink(itemLoc)
-    if IsCorruptedItem(itemLink) then
+function Module:ApplyIcon(button, itemLink)
+    if itemLink and IsCorruptedItem(itemLink) then
         local _, icon = Scanner:GetCorruptionByItemLink(itemLink)
         CreateCorruptionIcon(button, icon)
         AddNzothIconOverlay(button)
-    end
-end
-
-function CT_AdiBags_UpdateAfter()
-    if debounce ~= true then
-        debounce = true
-        C_Timer.After(.5, function()
-            debounce = false
-            for i=1,600 do
-                local frame = _G["AdiBagsItemButton"..i]
-                if frame then
-                    C_Timer.After(.5, function() CT_AdiBagsItemButton_CIMIUpdateIcon(frame) end)
-                end
-            end
-            for i=1,28 do
-                local frame = _G["AdiBagsBankItemButton"..i]
-                if frame then
-                    C_Timer.After(.5, function() CT_AdiBagsItemButton_CIMIUpdateIcon(frame) end)
-                end
-            end
-        end)
-    end
-end
-
-function CT_Bagnon_Update(button)
-    if not button then return end
-        local bag, slot = button:GetParent():GetID(), button:GetID()
-        -- need to catch 0, 0 and 100, 0 here because the bank frame doesn't
-        -- load everything immediately, so the OnUpdate needs to run until those frames are opened.
-        if (bag == nil) or (slot == nil) or (bag == 0 and slot == 0) or (bag == 100 and slot == 0) then return end
-        local itemLoc = ItemLocation:CreateFromBagAndSlot(bag, slot)
-
-        if (not itemLoc:IsValid()) then return end
-
-        local itemLink = C_Item.GetItemLink(itemLoc)
-        if IsCorruptedItem(itemLink) then
-            local _, icon = Scanner:GetCorruptionByItemLink(itemLink)
-            CreateCorruptionIcon(button, icon)
-            AddNzothIconOverlay(button)
-        end
-end
-
-function CT_Bagnon_UpdateBorder(button)
-    if not button.info.link then return end
-
-    local itemLink = button.info.link
-    if IsCorruptedItem(itemLink) then
-        button.IconOverlay:SetShown(true)
     end
 end